requirements.txt
torch>=2.3,<2.11
torchvision>=0.18,<0.26
numpy>=1.26,<3
Pillow>=10,<13
PyYAML>=6,<7
tqdm>=4.66,<5
scipy>=1.13,<1.17
transformers>=4.46,<5
